Fix exit path in the splash daemon.
From: Daniel Drake <dsd@gentoo.org>

We currently have an exit race: exit_cmd() sends SIGTERM to the sighandler
thread, then both sighandler and exit_cmd() call exit().

This adds another signal to the signal handler (SIGINT) which is equivalent
to SIGTERM except it means "another thread wants me to go away" as opposed
to "someone wants me to terminate everything".

cmd_exit() now sends SIGINT and waits for the signal handler to finish up
and go away before cleaning everything up.

I also added a VT_RELDISP ioctl to the cleanup path just to be 100% sure
that the kernel knows we have released the terminal. I'm investigating a race
where the kernel asks us to release the terminal, but as we are shutting
down we don't have time to process the signal properly. This isn't a complete
fix but should be safe anyway.