gentoo: fix handling of scheduled services
The problem is similar to the inactive-->starting case -- those services
were being counted twice, and the 2nd time was causing the progress bar
to step backwards.

This time, the step backwards was caused by double counting of the
scheduled services, which cause the progress bar to increment when
they initially get scheduled (due to depending on a service that went
inactive), and then again if they go from scheduled to starting while
the splash is still active.

In this case, the start_out hook gets called with scheduled=1 inactive=0
wasinactive=0 starting=1 started=0 so we can just ignore those cases.