core: make it possible for the splash to work in unsupported video modes
If there is no config for the current video mode, the splash daemon will
try to use the config file for a lower resolution.  If it succeeds, the
picture is displayed in the middle of the screen and there is a black
border around it.

This commit also removes support for exporting the background buffer to
a file.